==> 00_read_me.ipynb <==
## 01_general_data

##### This notebook takes some information from multiple raw data files to create a dataframe with the following columns:
##### -Season - the NCAA basketball season in the form of calendar year
##### -Seed - the team seed in the NCAA tournament, ranging from 1 to 16
##### -TeamID - a unique identifier of each team, which is included in each raw data set
##### -RankingDayNum - the day number after which season the begins. Only data for Day 133 is retained in the dataframe, which corresponds to the day before the tournament begins.
##### -SystemName - the abbreviation of the name of the ranking system
##### -OrdinalRank - the team rank assigned by the corresponding ranking system
##### -Wins - regular season wins
##### -Losses - regular season losses
##### -Win% - regular season winning percentage
##### -T_Wins - number of tournament wins, ranging from 0-6


### Raw data files imported

##### MTeams.csv
##### MMasseyOrdinals.csv
##### MNCAATourneySeeds.csv
##### MRegularSeasonCompactResults.csv
##### MNCAATourneyCompactResults.csv

### Dataframes imported

##### n/a

### Dataframes exported

#### team_data.csv

### 02_regular_season_data_clean

##### This notebook takes data from a raw data file in the form of every regular season game played. The basketball stats are aggregated to calculate season-long stats. These season-long stats are compiled in a dataframe with the following columns:
##### -Season - the NCAA basketball season in the form of calendar year
##### -TeamID - a unique identifier of each team, which is included in each raw data set.

##### -FGM - Field Goals Made 
##### -FGM_Ag - Field Goals Against 
##### -FGM_pg - Field Goals Made per game
##### -FGM_Ag_pg - Field Goals Made Against per game
##### -FGA - Field Goals Attempted 
##### -FGA_Ag - Field Goals Attempted Against 
##### -FG% - Field Goal Percentage 
##### -FG%_Ag - Field Goal Percentage Against

##### -FGM3 - 3-pointers Made 
##### -FGM3_Ag - 3-pointers Against 
##### -FGM3_pg - 3-pointers per game
##### -FGM3_Ag_pg - 3-pointers Against per game
##### -FGA3 - 3-pointers Attempted 
##### -FGA3_Ag - 3-pointers Attempted Against 
##### -FG3% - 3-pointer Percentage 
##### -FG3%_Ag - 3-pointer Percentage Against

##### -FTM - Free Throws Made 
##### -FTM_Ag - Free Throws Against 
##### -FTM_pg - Free Throws per game
##### -FTM_Ag_pg - Free Throws Against per game
##### -FTA - Free Throws Attempted 
##### -FTA_Ag - Free Throws Attempted Against 
##### -FTA_pg - Free Throws Attempted per game
##### -FTA_Ag-pg - Free Throws Attempted Against per game
##### -FT% - Free Throws Percentage 

##### -OR - Offensive Rebounds 
##### -OR_Ag - Offensive Rebounds Against 
##### -OR_pg - Offensive Rebounds per game
##### -OR_Ag_pg - Offensive Rebounds Against per game

##### -DR - Defensive Rebounds 
##### -DR_Ag - Defensive Rebounds Against 
##### -DR_pg - Defensive Rebounds per game
##### -DR_Ag_pg - Defensive Rebounds Against per game

##### -Ast - Assists
##### -Ast_Ag - Assists Against 
##### -Ast_pg - Assists per game
##### -Ast_Ag_pg - Assists Against per game

##### -TO - Turnovers 
##### -TO_Ag - Turnovers Against 
##### -TO_pg - Turnovers per game
##### -TO_Ag_pg - Turnovers Against per game

##### -Stl - Steals
##### -Stl_Ag - Steals Against 
##### -Stl_pg - Steals per game
##### -Stl_Ag_pg - Steals Against per game

##### -Blk - Blocks
##### -Blk_Ag - Blocks Against 
##### -Blk_pg - Blocks per game
##### -Blk_Ag_pg - Blocks Against per game

##### -PF - Personal Fouls
##### -PF_Ag - Personal Fouls Against 
##### -PF_pg - Personal Fouls per game
##### -PF_Ag_pg - Personal Fouls Against per game

### Raw data files imported

#### MRegularSeasonDetailedResults.csv

### Dataframes imported

##### n/a

### Dataframes exported

#### season_stats.csv

## 03_rankings_FE

##### This notebook is used to determine which ranking systems to include in the final model. The chosen ranking systems are included in separate columns.
##### In addition to the data from the team_data dataframe from notebook 01_general_data, with the following exceptions:
#### the following columns are included in the created dataframe.
##### - Rank_MOR - The ordinal rank of the Moore rankings system
##### - Rank_POM - The ordinal rank of the Pomeroy rankings system
##### - Rank_SAG - The ordinal rank of the Sagarin rankings system
##### - Rank_COL - The ordinal rank of the Colley rankings system
##### - Rank_DOL - The ordinal rank of the Dolphin rankings system
##### - Rank_WLK - The ordinal rank of the MOR rankings system
#### the following columns are dropped from the created dataframe.
##### -RankingDayNum - the day number after which season the begins. Only data for Day 133 is retained in the dataframe, which corresponds to the day before the tournament begins.
##### -SystemName - the abbreviation of the name of the ranking system
##### -OrdinalRank - the team rank assigned by the corresponding ranking system

### Raw data files imported

#### n/a

### Dataframes imported

#### team_data.csv

### Dataframes exported

#### team_data_clean.csv

## 04_final_data_frame

##### In this notebook, the two dataframes created in the previous notebooks (team_data_clean from notebook 03_rankings_FE and season_stats from notebook 02_regular_season_data_clean).
##### All columns from these two dataframes are retained.

### Raw data files imported

#### n/a

### Dataframes imported

#### team_data_clean.csv
#### season_stats.csv

### Dataframes exported

#### final_df.csv

## 05_EDA

##### This notebook contains exploratory data analysis. 

## 06_model_final

##### In this notebook, an appropriate classification model type (Random Forest) is selected.
##### the final_df.csv dataframe is imported from notebook 04_final_data_frame.
##### The data is normalized and scaled and oversampled due to class imbalance.
##### The model is evaluated compared to historical NCAA bracketology data
