==> 01_general_data.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import scipy.stats as stats

##### Create dataframe of Team Names and Team IDs. The TeamID column will be used to join the other data tables into a single dataframe.

teams = pd.read_csv("data/MTeams.csv")
teams.drop(['FirstD1Season', 'LastD1Season'], axis=1, inplace=True)
teams

#### Import data table that shows the ranking of each team according to various ranking systems at various points in the season. Only RankingDayNum = 133 will be used, which corresponds to the day before the tournament begins. The imported data includes a column called SystemName, which includes all NCAA college basketball ranking systems.

rankings = pd.read_csv("data/MMasseyOrdinals.csv")
rankings = rankings[rankings['RankingDayNum'] == 133]
#rankings

#### Merge the team table with the rankings table to get the team name with the rankings

team_data = teams.merge(rankings, how='inner', on='TeamID')

#### Import data table that shows the seed number of each team of each tournament. Remove all leading or trailing letters (which denote conference) from the Seed column and convert to integer

seeds = pd.read_csv("data/MNCAATourneySeeds.csv")
seeds.dropna()
seeds['Seed'] = seeds['Seed'].str[1:]
seeds['Seed'] = seeds['Seed'].str.lstrip('0')
seeds['Seed'] = seeds['Seed'].str.rstrip('abcd')
seeds['Seed'] = seeds['Seed'].astype(int)
seeds

#### Merge team data table with seeds table on TeamID and Season. This will eliminate all of the rows of teams that did not make the NCAA tournament in a given year.

team_data = seeds.merge(team_data, how='inner', on=['TeamID', 'Season'])

#### Import data that includes every regular season game, which includes the TeamID (WTeamID, LTeamID) and Score (WScore and LScore) of the winning and losing team of each game. The total count that each team won and loss will be calculated for each season. This is used to get regular season win-loss record.

record = pd.read_csv("data/MRegularSeasonCompactResults.csv")
record = record.drop(['DayNum', 'WScore', 'LScore', 'WLoc', 'NumOT'], axis=1)
wins= record.groupby(['Season', 'WTeamID']).count()
wins = wins.rename(columns={'WTeamID':'TeamID', 'LTeamID': 'Wins'})
losses = record.groupby(['Season', 'LTeamID']).count()
losses = losses.rename(columns={'LTeamID':'TeamID', 'WTeamID': 'Losses'})
wins=wins['Wins']
losses=losses['Losses']

#### Merge number of wins and number of losses to the main team data df. Merge on TeamID and Season.

team_data = pd.merge(team_data, wins, left_on=['TeamID', 'Season'], right_on=['WTeamID', 'Season'])
team_data = pd.merge(team_data, losses, left_on=['TeamID', 'Season'], right_on=['LTeamID', 'Season'])

#### Create Win% Column

team_data["Win%"]=round(team_data["Wins"]/(team_data["Wins"]+team_data["Losses"])*100,4)

#### Import data, which includes every tournament game, which includes the TeamID (WTeamID, LTeamID) and Score (WScore and LScore) of the winning and losing team of each game. The total count that each team won will be calculated for each season. The number of wins is the target feature and will be added to the main team data dataframe.

tourney= pd.read_csv("data/MNCAATourneyCompactResults.csv")
tourney = tourney.drop(['DayNum', 'WScore', 'LScore', 'WLoc', 'NumOT'], axis=1)
t_wins= tourney.groupby(['Season', 'WTeamID']).count()
t_wins = t_wins.rename(columns={'WTeamID':'TeamID', 'LTeamID': 'T_Wins'})
t_wins=t_wins['T_Wins']

#### Merge number of tournament wins to df. This will be the target feature

team_data = pd.merge(team_data, t_wins, how='left', left_on=['TeamID', 'Season'], right_on=['WTeamID', 'Season'])


#### Replace NaN with 0's for teams without a win (Teams that lost in the first round)

team_data["T_Wins"] = team_data["T_Wins"].replace(np.nan, 0)

#### Export team data dataframe to csv. More features will be added to this dataframe in another notebook.

team_data.to_csv('team_data.csv')

team_data
