==> 02_regular_season_data_clean.ipynb <==
import numpy as np
import pandas as pd

#### Import detailed regular season data. This includes every regular season game that every team played in a give season. It includes the scores, as well as basketball statistics such as three-pointers, free throws, turnovers, etc. of both the winning and losing teams.

season_stats= pd.read_csv("data/MRegularSeasonDetailedResults.csv")
season_stats = season_stats.drop(['DayNum', 'WLoc', 'NumOT'], axis=1)

season_stats

### Calculate Games Played. This will be used to generate 'per game' basketball statistics

GP_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WFGM'].count())
GP_W.reset_index(inplace=True)
GP_W.rename(columns={'WTeamID': 'TeamID', 'WFGM':'GP_W'}, inplace=True)
GP_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LFGM'].count())
GP_L.reset_index(inplace=True)
GP_L.rename(columns={'LTeamID': 'TeamID', 'LFGM':'GP_L'}, inplace=True)
GP = pd.merge(left=GP_W, right=GP_L, how='outer', on=['Season', 'TeamID'])
GP.replace(np.nan, 0, inplace=True)
GP['GP'] = GP['GP_W'] + GP['GP_L']
GP = GP.drop(['GP_W', 'GP_L'], axis=1)
GP

#### Sum the total of each team per season. This is done by grouping each basketball statistic by Season and TeamID. This aggreagtion is done twice, once for the games a team won, and another for games the same team lossed.

#### A similar aggregation will be done to calculate 'allowed' or 'against' basketball statistics. These are denoted by "_Ag". For example, the feature "FGM_Ag" represents how many field goals a team allowed in a give season.

### Field Goals

Field Goals Made

FGM_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WFGM'].sum())
FGM_W.reset_index(inplace=True)
FGM_W.rename(columns={'WTeamID':'TeamID'}, inplace=True)
FGM_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LFGM'].sum())
FGM_L.reset_index(inplace=True)
FGM_L.rename(columns={'LTeamID':'TeamID'}, inplace=True)
FGM_tot = pd.merge(left=FGM_W, right=FGM_L, how='outer', on=['Season', 'TeamID'])
FGM_tot.fillna(0, inplace=True)
FGM_tot['FGM'] = FGM_tot['WFGM'] + FGM_tot['LFGM']
FGM_tot.drop(['WFGM', 'LFGM'], axis = 1, inplace=True)

#FGM = pd.merge(left=GP, right=FGM_tot, how='outer', on=['Season', 'TeamID'])
#FGM

Field Goals Attempted

FGA_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WFGA'].sum())
FGA_W.reset_index(inplace=True)
FGA_W.rename(columns={'WTeamID':'TeamID'}, inplace=True)
FGA_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LFGA'].sum())
FGA_L.reset_index(inplace=True)
FGA_L.rename(columns={'LTeamID':'TeamID'}, inplace=True)
FGA_tot = pd.merge(left=FGA_W, right=FGA_L, how='outer', on=['Season', 'TeamID'])
FGA_tot.fillna(0, inplace=True)
FGA_tot['FGA'] = FGA_tot['WFGA'] + FGA_tot['LFGA']
FGA_tot.drop(['WFGA', 'LFGA'], axis = 1, inplace=True)

Field Goal Percentage

FGP = pd.merge(left=FGM_tot, right=FGA_tot, how='outer', on=['Season', 'TeamID'])
FGP['FG%'] = round(FGP['FGM'] / FGP['FGA']*100,4)

Field Goal Made Against

FGM_W_Ag = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['LFGM'].sum())
FGM_W_Ag.reset_index(inplace=True)
FGM_W_Ag.rename(columns={'WTeamID':'TeamID'}, inplace=True)
FGM_L_Ag = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['WFGM'].sum())
FGM_L_Ag.reset_index(inplace=True)
FGM_L_Ag.rename(columns={'LTeamID':'TeamID'}, inplace=True)
FGM_tot_Ag = pd.merge(left=FGM_W_Ag, right=FGM_L_Ag, how='outer', on=['Season', 'TeamID'])
FGM_tot_Ag.fillna(0, inplace=True)
FGM_tot_Ag['FGM_Ag'] = FGM_tot_Ag['WFGM'] + FGM_tot_Ag['LFGM']
FGM_tot_Ag.drop(['WFGM', 'LFGM'], axis = 1, inplace=True)

Field Goal Attempts Against

FGA_W_Ag = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['LFGA'].sum())
FGA_W_Ag.reset_index(inplace=True)
FGA_W_Ag.rename(columns={'WTeamID':'TeamID'}, inplace=True)
FGA_L_Ag = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['WFGA'].sum())
FGA_L_Ag.reset_index(inplace=True)
FGA_L_Ag.rename(columns={'LTeamID':'TeamID'}, inplace=True)
FGA_tot_Ag = pd.merge(left=FGA_W_Ag, right=FGA_L_Ag, how='outer', on=['Season', 'TeamID'])
FGA_tot_Ag.fillna(0, inplace=True)
FGA_tot_Ag['FGA_Ag'] = FGA_tot_Ag['WFGA'] + FGA_tot_Ag['LFGA']
FGA_tot_Ag.drop(['WFGA', 'LFGA'], axis = 1, inplace=True)

Field Goals Against Percentage

FGP_Ag = pd.merge(left=FGM_tot_Ag, right=FGA_tot_Ag, how='outer', on=['Season', 'TeamID'])
FGP_Ag['FG%_Ag'] = round(FGP_Ag['FGM_Ag'] / FGP_Ag['FGA_Ag']*100,4)

Field Goals Made / Against per game

df_FG = FGP.merge(FGP_Ag,how='outer', on=['Season', 'TeamID'])
df_FG = df_FG.merge(GP,how='outer', on=['Season', 'TeamID'])
df_FG['FGM_pg'] = round(df_FG['FGM']/df_FG['GP'],3)
df_FG['FGM_Ag_pg'] = round(df_FG['FGM_Ag']/df_FG['GP'],3)
df_FG.drop('GP', axis=1, inplace=True)
df_FG

### 3-Pointers

3-Pointers Made

FGM3_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WFGM3'].sum())
FGM3_W.reset_index(inplace=True)
FGM3_W.rename(columns={'WTeamID':'TeamID'}, inplace=True)
FGM3_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LFGM3'].sum())
FGM3_L.reset_index(inplace=True)
FGM3_L.rename(columns={'LTeamID':'TeamID'}, inplace=True)
FGM3_tot = pd.merge(left=FGM3_W, right=FGM3_L, how='outer', on=['Season', 'TeamID'])
FGM3_tot.fillna(0, inplace=True)
FGM3_tot['FGM3'] = FGM3_tot['WFGM3'] + FGM3_tot['LFGM3']
FGM3_tot.drop(['WFGM3', 'LFGM3'], axis = 1, inplace=True)

3-Pointers Attempted

FGA3_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WFGA3'].sum())
FGA3_W.reset_index(inplace=True)
FGA3_W.rename(columns={'WTeamID':'TeamID'}, inplace=True)
FGA3_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LFGA3'].sum())
FGA3_L.reset_index(inplace=True)
FGA3_L.rename(columns={'LTeamID':'TeamID'}, inplace=True)
FGA3_tot = pd.merge(left=FGA3_W, right=FGA3_L, how='outer', on=['Season', 'TeamID'])
FGA3_tot.fillna(0, inplace=True)
FGA3_tot['FGA3'] = FGA3_tot['WFGA3'] + FGA3_tot['LFGA3']
FGA3_tot.drop(['WFGA3', 'LFGA3'], axis = 1, inplace=True)

3-Pointer Percentage

FG3P = pd.merge(left=FGM3_tot, right=FGA3_tot, how='outer', on=['Season', 'TeamID'])
FG3P['FG3%'] = round(FG3P['FGM3'] / FG3P['FGA3']*100,4)

3-Pointers Made Against

FGM3_W_Ag = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['LFGM3'].sum())
FGM3_W_Ag.reset_index(inplace=True)
FGM3_W_Ag.rename(columns={'WTeamID':'TeamID'}, inplace=True)
FGM3_L_Ag = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['WFGM3'].sum())
FGM3_L_Ag.reset_index(inplace=True)
FGM3_L_Ag.rename(columns={'LTeamID':'TeamID'}, inplace=True)
FGM3_tot_Ag = pd.merge(left=FGM3_W_Ag, right=FGM3_L_Ag, how='outer', on=['Season', 'TeamID'])
FGM3_tot_Ag.fillna(0, inplace=True)
FGM3_tot_Ag['FGM3_Ag'] = FGM3_tot_Ag['WFGM3'] + FGM3_tot_Ag['LFGM3']
FGM3_tot_Ag.drop(['WFGM3', 'LFGM3'], axis = 1, inplace=True)

3-Pointers Attempted Against

FGA3_W_Ag = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['LFGA3'].sum())
FGA3_W_Ag.reset_index(inplace=True)
FGA3_W_Ag.rename(columns={'WTeamID':'TeamID'}, inplace=True)
FGA3_L_Ag = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['WFGA3'].sum())
FGA3_L_Ag.reset_index(inplace=True)
FGA3_L_Ag.rename(columns={'LTeamID':'TeamID'}, inplace=True)
FGA3_tot_Ag = pd.merge(left=FGA3_W_Ag, right=FGA3_L_Ag, how='outer', on=['Season', 'TeamID'])
FGA3_tot_Ag.fillna(0, inplace=True)
FGA3_tot_Ag['FGA3_Ag'] = FGA3_tot_Ag['WFGA3'] + FGA3_tot_Ag['LFGA3']
FGA3_tot_Ag.drop(['WFGA3', 'LFGA3'], axis = 1, inplace=True)

3-Pointers Percentage Against

FG3P_Ag = pd.merge(left=FGM3_tot_Ag, right=FGA3_tot_Ag, how='outer', on=['Season', 'TeamID'])
FG3P_Ag['FG3%_Ag'] = round(FG3P_Ag['FGM3_Ag'] / FG3P_Ag['FGA3_Ag']*100,4)

3-Pointers Made / Against per game

df_FG3 = FG3P.merge(FG3P_Ag,how='outer', on=['Season', 'TeamID'])
df_FG3 = df_FG3.merge(GP,how='outer', on=['Season', 'TeamID'])
df_FG3['FGM3_pg'] = round(df_FG3['FGM3']/df_FG3['GP'],3)
df_FG3['FGM3_Ag_pg'] = round(df_FG3['FGM3_Ag']/df_FG3['GP'],3)
df_FG3.drop('GP', axis=1, inplace=True)
df_FG3

### Free Throws

Free Throws Made

FTM_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WFTM'].sum())
FTM_W.reset_index(inplace=True)
FTM_W.rename(columns={'WTeamID':'TeamID'}, inplace=True)
FTM_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LFTM'].sum())
FTM_L.reset_index(inplace=True)
FTM_L.rename(columns={'LTeamID':'TeamID'}, inplace=True)
FTM_tot = pd.merge(left=FTM_W, right=FTM_L, how='outer', on=['Season', 'TeamID'])
FTM_tot.fillna(0, inplace=True)
FTM_tot['FTM'] = FTM_tot['WFTM'] + FTM_tot['LFTM']
FTM_tot.drop(['WFTM', 'LFTM'], axis = 1, inplace=True)

Free Throws Attempted

FTA_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WFTA'].sum())
FTA_W.reset_index(inplace=True)
FTA_W.rename(columns={'WTeamID':'TeamID'}, inplace=True)
FTA_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LFTA'].sum())
FTA_L.reset_index(inplace=True)
FTA_L.rename(columns={'LTeamID':'TeamID'}, inplace=True)
FTA_tot = pd.merge(left=FTA_W, right=FTA_L, how='outer', on=['Season', 'TeamID'])
FTA_tot.fillna(0, inplace=True)
FTA_tot['FTA'] = FTA_tot['WFTA'] + FTA_tot['LFTA']
FTA_tot.drop(['WFTA', 'LFTA'], axis = 1, inplace=True)

Free Throw Percentage

FTP = pd.merge(left=FTM_tot, right=FTA_tot, how='outer', on=['Season', 'TeamID'])
FTP['FT%'] = round(FTP['FTM'] / FTP['FTA']*100,4)

Free Throws Made Against

FTM_W_Ag = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['LFTM'].sum())
FTM_W_Ag.reset_index(inplace=True)
FTM_W_Ag.rename(columns={'WTeamID':'TeamID'}, inplace=True)
FTM_L_Ag = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['WFTM'].sum())
FTM_L_Ag.reset_index(inplace=True)
FTM_L_Ag.rename(columns={'LTeamID':'TeamID'}, inplace=True)
FTM_tot_Ag = pd.merge(left=FTM_W_Ag, right=FTM_L_Ag, how='outer', on=['Season', 'TeamID'])
FTM_tot_Ag.fillna(0, inplace=True)
FTM_tot_Ag['FTM_Ag'] = FTM_tot_Ag['WFTM'] + FTM_tot_Ag['LFTM']
FTM_tot_Ag.drop(['WFTM', 'LFTM'], axis = 1, inplace=True)

Free Throws Attempted Against

FTA_W_Ag = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['LFTA'].sum())
FTA_W_Ag.reset_index(inplace=True)
FTA_W_Ag.rename(columns={'WTeamID':'TeamID'}, inplace=True)
FTA_L_Ag = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['WFTA'].sum())
FTA_L_Ag.reset_index(inplace=True)
FTA_L_Ag.rename(columns={'LTeamID':'TeamID'}, inplace=True)
FTA_tot_Ag = pd.merge(left=FTA_W_Ag, right=FTA_L_Ag, how='outer', on=['Season', 'TeamID'])
FTA_tot_Ag.fillna(0, inplace=True)
FTA_tot_Ag['FTA_Ag'] = FTA_tot_Ag['WFTA'] + FTA_tot_Ag['LFTA']
FTA_tot_Ag.drop(['WFTA', 'LFTA'], axis = 1, inplace=True)

Free Throws Made / Attempted per game

FT_Ag_combined = pd.merge(left=FTM_tot_Ag, right=FTA_tot_Ag, how='outer', on=['Season', 'TeamID'])
FT_combined = pd.merge(left=FTP, right=FT_Ag_combined, how='outer', on=['Season', 'TeamID'])

df_FT = FT_combined.merge(GP,how='outer', on=['Season', 'TeamID'])
df_FT['FTM_pg'] = round(df_FT['FTM']/df_FT['GP'],3)
df_FT['FTM_Ag_pg'] = round(df_FT['FTM_Ag']/df_FT['GP'],3)
df_FT.drop('GP', axis=1, inplace=True)
df_FT

### Offensive Rebounds

OR_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WOR'].sum())
OR_W.reset_index(inplace=True)
OR_W.rename(columns={'WTeamID':'TeamID'}, inplace=True)
OR_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LOR'].sum())
OR_L.reset_index(inplace=True)
OR_L.rename(columns={'LTeamID':'TeamID'}, inplace=True)
OR_tot = pd.merge(left=OR_W, right=OR_L, how='outer', on=['Season', 'TeamID'])
OR_tot.fillna(0, inplace=True)
OR_tot['OR'] = OR_tot['WOR'] + OR_tot['LOR']
OR_tot.drop(['WOR', 'LOR'], axis = 1, inplace=True)

Offensive Rebounds Against

OR_W_Ag = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['LOR'].sum())
OR_W_Ag.reset_index(inplace=True)
OR_W_Ag.rename(columns={'WTeamID':'TeamID'}, inplace=True)
OR_L_Ag = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['WOR'].sum())
OR_L_Ag.reset_index(inplace=True)
OR_L_Ag.rename(columns={'LTeamID':'TeamID'}, inplace=True)
OR_tot_Ag = pd.merge(left=OR_W_Ag, right=OR_L_Ag, how='outer', on=['Season', 'TeamID'])
OR_tot_Ag.fillna(0, inplace=True)
OR_tot_Ag['OR_Ag'] = OR_tot_Ag['WOR'] + OR_tot_Ag['LOR']
OR_tot_Ag.drop(['WOR', 'LOR'], axis = 1, inplace=True)

Offensive Rebounds per game

OR_combined = pd.merge(left=OR_tot, right=OR_tot_Ag, how='outer', on=['Season', 'TeamID'])

df_OR = OR_combined.merge(GP,how='outer', on=['Season', 'TeamID'])
df_OR['OR_pg'] = round(df_OR['OR']/df_OR['GP'],3)
df_OR['OR_Ag_pg'] = round(df_OR['OR_Ag']/df_OR['GP'],3)
df_OR.drop('GP', axis=1, inplace=True)
df_OR

### Defensive Rebounds

DR_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WDR'].sum())
DR_W.reset_index(inplace=True)
DR_W.rename(columns={'WTeamID':'TeamID'}, inplace=True)
DR_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LDR'].sum())
DR_L.reset_index(inplace=True)
DR_L.rename(columns={'LTeamID':'TeamID'}, inplace=True)
DR_tot = pd.merge(left=DR_W, right=DR_L, how='outer', on=['Season', 'TeamID'])
DR_tot.fillna(0, inplace=True)
DR_tot['DR'] = DR_tot['WDR'] + DR_tot['LDR']
DR_tot.drop(['WDR', 'LDR'], axis = 1, inplace=True)

Defensive Rebounds Against

DR_W_Ag = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['LDR'].sum())
DR_W_Ag.reset_index(inplace=True)
DR_W_Ag.rename(columns={'WTeamID':'TeamID'}, inplace=True)
DR_L_Ag = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['WDR'].sum())
DR_L_Ag.reset_index(inplace=True)
DR_L_Ag.rename(columns={'LTeamID':'TeamID'}, inplace=True)
DR_tot_Ag = pd.merge(left=DR_W_Ag, right=DR_L_Ag, how='outer', on=['Season', 'TeamID'])
DR_tot_Ag.fillna(0, inplace=True)
DR_tot_Ag['DR_Ag'] = DR_tot_Ag['WDR'] + DR_tot_Ag['LDR']
DR_tot_Ag.drop(['WDR', 'LDR'], axis = 1, inplace=True)

Defensive Rebounds per game

DR_combined = pd.merge(left=DR_tot, right=DR_tot_Ag, how='outer', on=['Season', 'TeamID'])

df_DR = DR_combined.merge(GP,how='outer', on=['Season', 'TeamID'])
df_DR['DR_pg'] = round(df_DR['DR']/df_DR['GP'],3)
df_DR['DR_Ag_pg'] = round(df_DR['DR_Ag']/df_DR['GP'],3)
df_DR.drop('GP', axis=1, inplace=True)
df_DR

### Assists

Ast_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WAst'].sum())
Ast_W.reset_index(inplace=True)
Ast_W.rename(columns={'WTeamID':'TeamID'}, inplace=True)
Ast_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LAst'].sum())
Ast_L.reset_index(inplace=True)
Ast_L.rename(columns={'LTeamID':'TeamID'}, inplace=True)
Ast_tot = pd.merge(left=Ast_W, right=Ast_L, how='outer', on=['Season', 'TeamID'])
Ast_tot.fillna(0, inplace=True)
Ast_tot['Ast'] = Ast_tot['WAst'] + Ast_tot['LAst']
Ast_tot.drop(['WAst', 'LAst'], axis = 1, inplace=True)

Assists against

Ast_W_Ag = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['LAst'].sum())
Ast_W_Ag.reset_index(inplace=True)
Ast_W_Ag.rename(columns={'WTeamID':'TeamID'}, inplace=True)
Ast_L_Ag = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['WAst'].sum())
Ast_L_Ag.reset_index(inplace=True)
Ast_L_Ag.rename(columns={'LTeamID':'TeamID'}, inplace=True)
Ast_tot_Ag = pd.merge(left=Ast_W_Ag, right=Ast_L_Ag, how='outer', on=['Season', 'TeamID'])
Ast_tot_Ag.fillna(0, inplace=True)
Ast_tot_Ag['Ast_Ag'] = Ast_tot_Ag['WAst'] + Ast_tot_Ag['LAst']
Ast_tot_Ag.drop(['WAst', 'LAst'], axis = 1, inplace=True)

Assists per game

Ast_combined = pd.merge(left=Ast_tot, right=Ast_tot_Ag, how='outer', on=['Season', 'TeamID'])

df_Ast = Ast_combined.merge(GP,how='outer', on=['Season', 'TeamID'])
df_Ast['Ast_pg'] = round(df_Ast['Ast']/df_Ast['GP'],3)
df_Ast['Ast_Ag_pg'] = round(df_Ast['Ast_Ag']/df_Ast['GP'],3)
df_Ast.drop('GP', axis=1, inplace=True)
df_Ast

### Turnovers

TO_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WTO'].sum())
TO_W.reset_index(inplace=True)
TO_W.rename(columns={'WTeamID':'TeamID'}, inplace=True)
TO_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LTO'].sum())
TO_L.reset_index(inplace=True)
TO_L.rename(columns={'LTeamID':'TeamID'}, inplace=True)
TO_tot = pd.merge(left=TO_W, right=TO_L, how='outer', on=['Season', 'TeamID'])
TO_tot.fillna(0, inplace=True)
TO_tot['TO'] = TO_tot['WTO'] + TO_tot['LTO']
TO_tot.drop(['WTO', 'LTO'], axis = 1, inplace=True)

Turnovers Against

TO_W_Ag = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['LTO'].sum())
TO_W_Ag.reset_index(inplace=True)
TO_W_Ag.rename(columns={'WTeamID':'TeamID'}, inplace=True)
TO_L_Ag = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['WTO'].sum())
TO_L_Ag.reset_index(inplace=True)
TO_L_Ag.rename(columns={'LTeamID':'TeamID'}, inplace=True)
TO_tot_Ag = pd.merge(left=TO_W_Ag, right=TO_L_Ag, how='outer', on=['Season', 'TeamID'])
TO_tot_Ag.fillna(0, inplace=True)
TO_tot_Ag['TO_Ag'] = TO_tot_Ag['WTO'] + TO_tot_Ag['LTO']
TO_tot_Ag.drop(['WTO', 'LTO'], axis = 1, inplace=True)

Turnovers per game 

TO_combined = pd.merge(left=TO_tot, right=TO_tot_Ag, how='outer', on=['Season', 'TeamID'])

df_TO = TO_combined.merge(GP,how='outer', on=['Season', 'TeamID'])
df_TO['TO_pg'] = round(df_TO['TO']/df_TO['GP'],3)
df_TO['TO_Ag_pg'] = round(df_TO['TO_Ag']/df_TO['GP'],3)
df_TO.drop('GP', axis=1, inplace=True)
df_TO

### Steals

Stl_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WStl'].sum())
Stl_W.reset_index(inplace=True)
Stl_W.rename(columns={'WTeamID':'TeamID'}, inplace=True)
Stl_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LStl'].sum())
Stl_L.reset_index(inplace=True)
Stl_L.rename(columns={'LTeamID':'TeamID'}, inplace=True)
Stl_tot = pd.merge(left=Stl_W, right=Stl_L, how='outer', on=['Season', 'TeamID'])
Stl_tot.fillna(0, inplace=True)
Stl_tot['Stl'] = Stl_tot['WStl'] + Stl_tot['LStl']
Stl_tot.drop(['WStl', 'LStl'], axis = 1, inplace=True)

Steals Against

Stl_W_Ag = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['LStl'].sum())
Stl_W_Ag.reset_index(inplace=True)
Stl_W_Ag.rename(columns={'WTeamID':'TeamID'}, inplace=True)
Stl_L_Ag = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['WStl'].sum())
Stl_L_Ag.reset_index(inplace=True)
Stl_L_Ag.rename(columns={'LTeamID':'TeamID'}, inplace=True)
Stl_tot_Ag = pd.merge(left=Stl_W_Ag, right=Stl_L_Ag, how='outer', on=['Season', 'TeamID'])
Stl_tot_Ag.fillna(0, inplace=True)
Stl_tot_Ag['Stl_Ag'] = Stl_tot_Ag['WStl'] + Stl_tot_Ag['LStl']
Stl_tot_Ag.drop(['WStl', 'LStl'], axis = 1, inplace=True)

Steal per game

Stl_combined = pd.merge(left=Stl_tot, right=Stl_tot_Ag, how='outer', on=['Season', 'TeamID'])

df_Stl = Stl_combined.merge(GP,how='outer', on=['Season', 'TeamID'])
df_Stl['Stl_pg'] = round(df_Stl['Stl']/df_Stl['GP'],3)
df_Stl['Stl_Ag_pg'] = round(df_Stl['Stl_Ag']/df_Stl['GP'],3)
df_Stl.drop('GP', axis=1, inplace=True)
df_Stl

### Blocks

Blk_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WBlk'].sum())
Blk_W.reset_index(inplace=True)
Blk_W.rename(columns={'WTeamID':'TeamID'}, inplace=True)
Blk_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LBlk'].sum())
Blk_L.reset_index(inplace=True)
Blk_L.rename(columns={'LTeamID':'TeamID'}, inplace=True)
Blk_tot = pd.merge(left=Blk_W, right=Blk_L, how='outer', on=['Season', 'TeamID'])
Blk_tot.fillna(0, inplace=True)
Blk_tot['Blk'] = Blk_tot['WBlk'] + Blk_tot['LBlk']
Blk_tot.drop(['WBlk', 'LBlk'], axis = 1, inplace=True)

Blocks Against

Blk_W_Ag = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['LBlk'].sum())
Blk_W_Ag.reset_index(inplace=True)
Blk_W_Ag.rename(columns={'WTeamID':'TeamID'}, inplace=True)
Blk_L_Ag = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['WBlk'].sum())
Blk_L_Ag.reset_index(inplace=True)
Blk_L_Ag.rename(columns={'LTeamID':'TeamID'}, inplace=True)
Blk_tot_Ag = pd.merge(left=Blk_W_Ag, right=Blk_L_Ag, how='outer', on=['Season', 'TeamID'])
Blk_tot_Ag.fillna(0, inplace=True)
Blk_tot_Ag['Blk_Ag'] = Blk_tot_Ag['WBlk'] + Blk_tot_Ag['LBlk']
Blk_tot_Ag.drop(['WBlk', 'LBlk'], axis = 1, inplace=True)

Blocks per game

Blk_combined = pd.merge(left=Blk_tot, right=Blk_tot_Ag, how='outer', on=['Season', 'TeamID'])

df_Blk = Blk_combined.merge(GP,how='outer', on=['Season', 'TeamID'])
df_Blk['Blk_pg'] = round(df_Blk['Blk']/df_Blk['GP'],3)
df_Blk['Blk_Ag_pg'] = round(df_Blk['Blk_Ag']/df_Blk['GP'],3)
df_Blk.drop('GP', axis=1, inplace=True)
df_Blk

### Personal Fouls

PF_W = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['WPF'].sum())
PF_W.reset_index(inplace=True)
PF_W.rename(columns={'WTeamID':'TeamID'}, inplace=True)
PF_L = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['LPF'].sum())
PF_L.reset_index(inplace=True)
PF_L.rename(columns={'LTeamID':'TeamID'}, inplace=True)
PF_tot = pd.merge(left=PF_W, right=PF_L, how='outer', on=['Season', 'TeamID'])
PF_tot.fillna(0, inplace=True)
PF_tot['PF'] = PF_tot['WPF'] + PF_tot['LPF']
PF_tot.drop(['WPF', 'LPF'], axis = 1, inplace=True)

Personal Fould Against

PF_W_Ag = pd.DataFrame(season_stats.groupby(['Season', 'WTeamID'])['LPF'].sum())
PF_W_Ag.reset_index(inplace=True)
PF_W_Ag.rename(columns={'WTeamID':'TeamID'}, inplace=True)
PF_L_Ag = pd.DataFrame(season_stats.groupby(['Season', 'LTeamID'])['WPF'].sum())
PF_L_Ag.reset_index(inplace=True)
PF_L_Ag.rename(columns={'LTeamID':'TeamID'}, inplace=True)
PF_tot_Ag = pd.merge(left=PF_W_Ag, right=PF_L_Ag, how='outer', on=['Season', 'TeamID'])
PF_tot_Ag.fillna(0, inplace=True)
PF_tot_Ag['PF_Ag'] = PF_tot_Ag['WPF'] + PF_tot_Ag['LPF']
PF_tot_Ag.drop(['WPF', 'LPF'], axis = 1, inplace=True)

Personal Fouls per game

PF_combined = pd.merge(left=PF_tot, right=PF_tot_Ag, how='outer', on=['Season', 'TeamID'])

df_PF = PF_combined.merge(GP,how='outer', on=['Season', 'TeamID'])
df_PF['PF_pg'] = round(df_PF['PF']/df_PF['GP'],3)
df_PF['PF_Ag_pg'] = round(df_PF['PF_Ag']/df_PF['GP'],3)
df_PF.drop('GP', axis=1, inplace=True)
df_PF

#### Combine all DataFrames to make a season statistics DataFrame

from functools import reduce

data_frames = [df_FG, df_FG3, df_FT, df_OR, df_DR, df_Ast, df_TO, df_Stl, df_Blk, df_PF]

df_season_stats = reduce(lambda  left,right: pd.merge(left,right,on=['Season', 'TeamID'], how='outer'), data_frames)



df_season_stats

#### Export season statistics dataframe to csv. This will be combined with the team data dataframe in another notebook. More features will be added to this dataframe in another notebook.

df_season_stats.to_csv('season_stats.csv')
