==> 04_final_data_frame.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import scipy.stats as stats

#### In this notebook, the cleaned team data (team_data_clean) from notebook 03_ranking is combined with the regular season stats (season_stats) to create the final dataframe for use in the classifcation model.

team_data_clean = pd.read_csv("team_data_clean.csv")
season_stats= pd.read_csv("season_stats.csv")

team_data_clean = team_data_clean.drop("Unnamed: 0", axis=1)
team_data_clean

season_stats= season_stats.drop("Unnamed: 0", axis=1)
season_stats

df = team_data_clean.merge(season_stats, how='inner', on=['TeamID', 'Season'])
df

df.to_csv('final_df.csv')
