==> 06_model_final.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
#import scipy.stats as stats
from sklearn import metrics

from sklearn.model_selection import train_test_split
from sklearn.model_selection import cross_val_score
from sklearn.linear_model import LinearRegression, LogisticRegression
from sklearn.metrics import accuracy_score, confusion_matrix, classification_report
from sklearn import svm
from sklearn.tree import DecisionTreeClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import MultinomialNB

import warnings
warnings.filterwarnings('ignore')

#### The target feature (number of tournmanet wins) is highly imbalanced, as is expected. For every team that wins the championship (6 wins), there are 32 teams that lose in the first round (0 wins).

#### The data shows that the number of wins of a class is approximately half of the number of wins of the previous class. 

####    -If the data was 100% complete. The relative amount of entries in each class would be EXACTLY half of its preceeding.
####    - It is noted that the exception to the above statement is between the 5 wins and 6 wins classes. This is expected since there is no 7 wins class. While two teams earn 5 wins. One of these teams will also earn a 6th win. So with 100% complete data. The 5 wins class and 6 wins class would have EQUAL number of entries. 

##### Oversample the data to compensate for the imbalanced data. Use SMOTE and class_weights

from imblearn.over_sampling import SMOTE
smote = SMOTE()

from sklearn.utils import class_weight

final_df = pd.read_csv('final_df.csv')

final_df.drop(['Unnamed: 0', 'TeamName','TeamID'], axis=1, inplace=True)

final_df

##### Normalize and Scale the data

from sklearn import preprocessing
scaler = preprocessing.StandardScaler()

##### Calculate the weight of each class and convert to a dictionary.

cw = class_weight.compute_class_weight(class_weight='balanced', classes=np.unique(final_df['T_Wins']),y=final_df['T_Wins'])


categories=list(final_df['T_Wins'].sort_values().unique())
cw_list=list(cw)
weights=dict(zip(categories, cw_list))


#DEFINE Train set and Test set
X = final_df
X = scaler.fit_transform(X)
X = pd.DataFrame(X)
y = final_df['T_Wins']
#X,y = smote.fit_resample(X,y)
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.35,random_state=101)
X_train, y_train = smote.fit_resample(X_train,y_train)

##### Define Classificaiton models to consider.

models = [RandomForestClassifier(n_estimators=100, class_weight=weights), KNeighborsClassifier(n_neighbors=17), svm.SVC(class_weight=weights), LogisticRegression(class_weight=weights)]


#### Run different metrics on each model to determine the most suitable. 
#### Logisitic Regression performs the best but raises concerns of overfitted data.
#### Random Forest model yields the strongest results with reaosnable accuracy.

model_selection_rank = pd.DataFrame()
for model in models:
    model.fit(X_train, y_train)
    scores = cross_val_score(model, X_train, y_train)
    model_selection_rank[model.__str__()] = scores
    model_selection_rank   
    y_predict=model.predict(X_test)
    ac=accuracy_score(y_predict,y_test)
    cm=confusion_matrix(y_test, y_predict)
    cr=classification_report(y_test, y_predict)
    print(model)
    print(f'Train Score: {model.score(X_train, y_train)}')
    print(f'Test Score: {model.score(X_test, y_test)}')
    print(f'Accuracy: {ac}')
    print(f'Confusion Matrix: \n {cm}')
    print(f'Classification Report: \n {cr}')

#CHOSEN MODEL
model = RandomForestClassifier(n_estimators=100, class_weight=weights)

model.fit(X_train, y_train)
y_pred = model.predict(X_test)
metrics.accuracy_score(y_test, y_pred)

result = pd.DataFrame({'actual':y_test, 'predicted':y_pred})

result['actual points earned'] = round(2**(result['actual']-1)-.01)
result['predicted points earned'] = round(2**(result['predicted']-1)-.01)

result['difference in score'] = abs(result['actual points earned'] - result['predicted points earned'])

nums = [0,1,2,3,4,5,6]
dictionary_result={}
pred_list=[]
act_list=[]
pred_correct_list=[]
for num in nums:
   # result_pred = result[(result['difference in score'] == 0) & (result['actual'] == num)]['actual'].count()
    result_pred = result[result['predicted'] == num]['predicted'].count()
    result_act = result[result['actual'] == num]['actual'].count()
    result_pred_correct = result[(result['difference in score'] == 0) & (result['actual'] == num)]['actual'].count()
    pred_list.append(result_pred)
    act_list.append(result_act)
    pred_correct_list.append(result_pred_correct)
    #dict.update({f'pred{num}' : result_pred, f'act{num}' : result_act})
dictionary_result={'predicted': pred_list, 'actual': act_list, 'predicted correct': pred_correct_list}
pd.DataFrame.from_dict(dictionary_result)

result

model_correct=[]
model_incorrect=[]
for num in list(result['actual'].sort_values().unique()):
    model_correct.append(result[(result['actual']>=num) & (result['predicted']>=num)]['actual'].count())
    model_incorrect.append(result[(result['actual']>=num) & (result['predicted']<num)]['actual'].count())

model_ac = pd.DataFrame()
model_ac['Round'] = [0,1,2,3,4,5,6]
model_ac['Model Correct'] = model_correct
model_ac['Model Incorrect'] = model_incorrect
model_ac.drop(0, inplace=True)
model_ac['Model Accuracy'] = round(model_ac['Model Correct']/(model_ac['Model Correct']+model_ac['Model Incorrect']),3)
model_ac


best_seed_correct = [23.1, 10.4, 4.2, 1.3, 0.6, 0.2]
best_seed_incorrect = [32-23.1, 16-10.4, 8-4.2, 4-1.3, 2-0.6, 1-0.2]
model_ac['Best-Seed Correct']=best_seed_correct
model_ac['Best-Seed Incorrect']=best_seed_incorrect
model_ac['Best-Seed Accuracy'] = round(model_ac['Best-Seed Correct']/(model_ac['Best-Seed Correct']+model_ac['Best-Seed Incorrect']),3)
model_ac


model_ac['Model Points'] = round(model_ac['Model Accuracy']*32,1)
model_ac['Best-Seed Points'] = round(model_ac['Best-Seed Accuracy']*32,1)

model_ac

model_ac['Model Points'].sum()

model_ac['Best-Seed Points'].sum()
